+++ .ci-release.ps1
@@ -12,9 +12,9 @@ if (Test-Path "C:\inetpub\wwwroot\tmp-build.enpaul\.git") {
 if (Test-Path "C:\inetpub\wwwroot\tmp-build.enpaul\.gitlab-ci.yml") {
   Remove-Item -Path "C:\inetpub\wwwroot\tmp-build.enpaul\.gitlab-ci.yml" -Force
 }
-if (Test-Path "C:\inetpub\wwwroot\tmp-build.enpaul\deploy.ps1") {
-  Remove-Item -Path "C:\inetpub\wwwroot\tmp-build.enpaul\deploy.ps1" -Force
-}
+
+Remove-Item -Path "C:\inetpub\wwwroot\tmp-build.enpaul\*.ps1" -Force
+
 if (Test-Path "C:\inetpub\wwwroot\tmp-build.enpaul\.gitignore") {
   Remove-Item -Path "C:\inetpub\wwwroot\tmp-build.enpaul\.gitignore" -Force
 }
@@ -22,8 +22,14 @@ if (Test-Path "C:\inetpub\wwwroot\tmp-build.enpaul\.htaccess") {
   Remove-Item -Path "C:\inetpub\wwwroot\tmp-build.enpaul\.htaccess" -Force
 }
 
-Rename-Item -Path "C:\inetpub\wwwroot\enpaul" -NewName "enpaul.old" -Force
+if (Test-Path "C:\inetpub\wwwroot\enpaul")
+  Rename-Item -Path "C:\inetpub\wwwroot\enpaul" -NewName "enpaul.old" -Force
+}
+
 Rename-Item -Path "C:\inetpub\wwwroot\tmp-build.enpaul" -NewName "enpaul" -Force
 
-Remove-Item -path "C:\inetpub\wwwroot\enpaul.old" -Recurse -Force
+if (Test-Path "C:\inetpub\wwwroot\enpaul.old") {
+  Remove-Item -path "C:\inetpub\wwwroot\enpaul.old" -Recurse -Force
+}
+
 exit
